deploy/requirements.txt
# Deployment Script Dependencies
# Install with: pip install -r requirements.txt

# Google Cloud SDK
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
google-cloud-run>=0.9.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# CLI utilities
click>=8.1.7
rich>=13.7.0  # Beautiful terminal formatting
tabulate>=0.9.0

# HTTP requests
requests>=2.31.0

# Security
cryptography>=41.0.0
secrets-manager>=1.0.0

# Testing
pytest>=7.4.0
pytest-timeout>=2.2.0
